cnn_image/requirements.txt
# pip install -r requirements.txt
# pip show torch
# pip list
torch==2.6.0
torchvision==0.21.0
Pillow
